\name{summary.binned}
\alias{summary.binned}
\title{Suummarizing binned data}
\description{ \code{summary} method for class \code{binned}.  }
\usage{ \S3method{summary}{binned}(b) }
\arguments{
  \item{b}{an object of class \code{"binned"}, usually, a result of a
        call to \code{\link{binned}} or \code{\link{as.binned}}. }
}
